Build script that includes git-sha in used library
To make building this library more dependable and
trackable, make a small build script that
builds an artifact including the git-sha key
and installs it to a local evoko-app folder.

Its run from this projects root folder with the
evoko-home app base folder as its sole argument.

E g:
./evoko-build.sh /home/jonasl/code/Evoko-GetARoom

Verification:
* Make some changes to the react-native code.
Run the build file and run an app that includes
the newly built files. Verify that it now includes
the changed files.
* Verify that the git-sha included in the binary
release name is correct.