Android: Fix contentOffset in the vertical ScrollView
Summary:
ScrollView does not support contentOffset on android
(yet). Applying the patch in this PR:
https://github.com/facebook/react-native/pull/12502

But on the vertical component (ReactScrollView)
fixes this.

Test Plan:
Use this patch on an app with a vertical
ScrollView. Set contentOffset to something
not default (x: 0, y: 0) and verify that
content is scrolled on rendering.

Reviewers: meros

Reviewed By: meros

Subscribers: jonas

Maniphest Tasks: T42

Differential Revision: http://192.168.5.120.xip.io/D41