Android: Fix lineheight same as font text drop
Summary:
If the lineheight is set to the same as
the font-size, this might cause overflow
if the line-height is bigger.

React caters for this, but as it is done, it
pushes text to the bottom of the bounding box.

Make the text more centered.

Test Plan:
* Inspect the text in all scenes on android,
text is now more centered and isn't cropped.

Reviewers: meros

Reviewed By: meros

Subscribers: jonas

Maniphest Tasks: T33

Differential Revision: http://192.168.5.120.xip.io/D46